\docType{methods}
\name{r}
\alias{r}
\alias{r,Circle-method}
\alias{r<-}
\alias{r<-,Circle-method}
\title{Coordinates of the Circle object}
\description{
  These methods are accessors to the center and radius of a
  Circle object.
}
\details{
  The \code{r(object)} method returns the radius length of
  the circle as an \code{\link{SVGLength}} object.

  The \code{r(object) <- value} method sets the radius
  length of the circle.
}

