\name{Circle}
\alias{Circle}
\alias{Circle-class}
\title{Circle Definition (VIRTUAL)}
\description{
  This class is used to defined all object that have circle
  properties. This class can't be directly instantiate and
  must be derived.
}
\details{
  The circle class is implemented as an circle where both
  radius are equals
}
\seealso{
  \code{\link{Circle}} the parent class
}

