\name{CSS}
\alias{CSS}
\alias{CSS-class}
\title{CSS inline attribute class (VIRTUAL)}
\description{
  This class handles all style proterties that are defined
  using inline CSS instructions. Furthermore, this class
  also provides some useful accessors to deal with inline
  style getter and/or setter.
}
\details{
  This class is mostly used in a multiple inheritance
  scheme, where it provides the additional definitions and
  implementations to deal with style properties. Due to
  this constraints there is no initialization method but a
  single prototype field in the class defintion.
}

