\docType{methods}
\name{URL}
\alias{URL}
\alias{URL,Gradient-method}
\title{Gradient document URL}
\arguments{
  \item{object}{is an SVG Gradient object}
}
\value{
  the gradient URL
}
\description{
  When a gradient object is added to the definition element
  of an SVG document it gets an XML ID. The \code{URL}
  method uses this identifier to forge a unique URL. This
  URL is used to link the gradient with a given shape, mask
  operation ...
}

