\name{Layout}
\alias{Layout}
\alias{Layout-class}
\title{Layout root class (VIRTUAL)}
\description{
  This class is the root class for all document layout
  methods. This class can't be directly instantiate and
  must be derived.
}
\seealso{
  \code{\link{SVGNode}} and \code{\link{Rectangle}} classes
}

