\name{LinearGradient}
\alias{LinearGradient}
\alias{LinearGradient-class}
\title{SVG Linear Gradient class}
\description{
  This class is used to describe a linear gradient, as part
  of the SVG 1.1 specifications. This class inherited from
  the Gradient class.
}
\seealso{
  \code{\link{Gradient}} and \code{\link{Vector}}.
}

