\docType{methods}
\name{targetAttribute}
\alias{targetAttribute}
\alias{targetAttribute,MappingColors-method}
\alias{targetAttribute<-}
\alias{targetAttribute<-,MappingColors-method}
\title{Target Style Attribute Accessors}
\arguments{
  \item{object}{the mapping colors instance object}
}
\value{
  a CSS style attribute name
}
\description{
  Given the set of \emph{targets}, to apply the
  \code{MappingColors} operation one as to provide the name
  of the \emph{style} attribute to change. The methods
  described here allow to specify this attribute.
}
\details{
  The \code{targetAttribute(object)} method retrieves the
  current target style attribute value.

  The \code{targetAttribute(object) <- value} method sets
  the target style attribute of the \code{object} mapping
  instance
}

