\name{MappingLinks}
\alias{MappingLinks}
\alias{MappingLinks-class}
\title{Links Mapping Operation}
\description{
  This class allows to associate hyperlinks (as \emph{HTML
  HREF} references) to the target shapes.
}
\details{
  Used of the Mapping Links operation is straightforward.
  One has onnly to provide html urls as input values. You
  can also initialize an url pattern (as a standard
  \code{\link{sprintf}} format expression) that will be
  used to forge the output urls.
}

