\name{MappingMasks}
\alias{MappingMasks}
\alias{MappingMasks-class}
\title{Mask Mapping Operation}
\description{
  This class creates SVG Masks to hide, partially or
  completely, target shapes. Input values are used to
  defined the percentage of the shape to be hidden.
  Furthermore, an angle parameter can be used to set the
  direction of the mask.
}

