\name{MappingTooltips}
\alias{MappingTooltips}
\alias{MappingTooltips-class}
\title{Tooltips Mapping Operation}
\description{
  This class allows to associate tooltips with any
  graphical shape. Tooltips are implemented as javascripts
  that are embedded with the template. They react to both
  \emph{onmouseover} and \emph{onmouseout} events. Default
  tooltip contain only one line that is filled with the
  content of input values.
}

