\name{RadialGradient}
\alias{RadialGradient}
\alias{RadialGradient-class}
\title{SVG Radial Gradient class}
\description{
  This class is used to describe a radial gradient, as part
  of the SVG 1.1 specifications. This class inherited from
  the Gradient class.
}
\seealso{
  \code{\link{Gradient}} and \code{\link{Circle}}.
}

