\docType{methods}
\name{defaultSearchAttrName}
\alias{defaultSearchAttrName}
\alias{defaultSearchAttrName,SVG-method}
\alias{defaultSearchAttrName<-}
\alias{defaultSearchAttrName<-,SVG-method}
\title{SVG Default Search Attribute Name}
\arguments{
  \item{object}{is an SVG class instance}
}
\value{
  This method returns the default search attribute name as
  a character string
}
\description{
  These methods allows to get/set the default attribute
  name used by the SVG class accessors
}
\details{
  The \code{defaultSearchAttrName(object)} method returns
  the default search attribute name

  The \code{defaultSearchAttrName(object) <- value} method
  sets the default search attribute name of object to value
}

