\docType{methods}
\name{isChildren}
\alias{isChildren}
\alias{isChildren,SVG-method}
\title{Check is a node is part of an SVG object}
\arguments{
  \item{object}{is the SVG instance}

  \item{node}{is an \code{XMLInternalNode}}
}
\value{
  \code{TRUE} if \emph{node} is part of the SVG object and
  \code{FALSE} otherwise
}
\description{
  This method check if a given node (as an
  \code{XMLInternalNode}) is part (\emph{i.e.} a children)
  of an SVG document.
}

