\docType{methods}
\name{duplicateNode}
\alias{duplicateNode}
\alias{duplicateNode,SVG-method}
\title{Duplicate part of a SVG document}
\arguments{
  \item{object}{is the SVG instance}

  \item{node}{is the root of the sub-tree to duplicate}

  \item{prefix}{is a string used to re-factor elements
  \emph{ID}}
}
\description{
  This method duplicates a part of an SVG object (given its
  \emph{root} node).
}

