\docType{methods}
\name{show}
\alias{print.SVG}
\alias{print.SVG,SVG-method}
\alias{show}
\alias{show,SVG-method}
\title{Show method for SVG objects}
\arguments{
  \item{object}{is a \code{\link{SVG}} instance.}
}
\description{
  Display the SVG object by opening the document in a
  webpage using the default browser.
}

