\docType{methods}
\name{summary.SVG}
\alias{summary.SVG}
\alias{summary.SVG,SVG-method}
\title{SVG Summary}
\arguments{
  \item{object}{for which the summary is returned}
}
\value{
  a list of various statistics that includes xml,
  animations and javascript
}
\description{
  summary of the SVG content
}

