\docType{methods}
\name{isRelative}
\alias{isRelative}
\alias{isRelative,SVGLength-method}
\title{Check for relative length}
\arguments{
  \item{object}{is the SVGLength object}
}
\value{
  TRUE for a relative length value, FALSE otherwise
}
\description{
  This method checks if the length value is relative to
  another one or not.
}

