\docType{methods}
\name{uUser}
\alias{uUser}
\alias{uUser,SVGLength-method}
\title{Value in user units system accessor}
\arguments{
  \item{object}{is an SVGLength object}
}
\value{
  return a numeric unit value or \code{NA} for relative
  length
}
\description{
  This method allows to access to the length value in the
  user units system, Unless the length value is given
  relatively to another one (\emph{eg} 'em', 'ex' or '%'
  units systems)
}

