\docType{methods}
\name{print.SVGLine}
\alias{print.SVGLine}
\alias{print.SVGLine,SVGLine-method}
\alias{print.SVGRect}
\alias{print.SVGRect,SVGRect-method}
\title{Print an SVG object to the terminal}
\arguments{
  \item{x}{the SVG object to print}
}
\value{
  the SVG object invisibly
}
\description{
  These methods attempt to print SVG object to the terminal
  as text XML
}

