\docType{methods}
\name{id}
\alias{id}
\alias{id,SVGNode-method}
\alias{id<-}
\alias{id<-,SVGNode-method}
\title{XML ID accessors of an SVG node}
\arguments{
  \item{object}{an SVG node}
}
\value{
  a character string
}
\description{
  This method returns the XML ID of an SVG Node
}
\details{
  The \code{id(object)<-} method can be used to set an XML
  ID to an SVG Node (see the detail section)
}

