\name{SVGNode}
\alias{SVGNode}
\alias{SVGNode-class}
\title{SVG Node root class (VIRTUAL)}
\description{
  SVGNode is the root class for all SVG elements sub-class.
  Normally it should not be directly instantiate. All SVG
  related classes derived from this class. It also provides
  methods to convert the class into XML representation
  using the \code{\link{.xml}} method. This class can't be
  directly instantiate and must be derived.
}

