\name{SVGShape}
\alias{SVGShape}
\alias{SVGShape-class}
\title{SVG Shapes root class (VIRTUAL)}
\description{
  This class is the root class for all SVG shapes. It
  contains the specifications for the core attributes of
  the basic shapes (detailed in SVG 1.1 specification).
  This class can't be directly instantiate and must be
  derived.
}
\seealso{
  \link{SVGNode}
}

