\docType{methods}
\name{print.SVGTransform}
\alias{print.SVGTransform}
\title{Print an SVGTransform object to the terminal}
\arguments{
  \item{x}{the SVGTransform object to print}
}
\value{
  the SVGTransform object invisibly
}
\description{
  This method attempts to print SVGTransform object to the
  terminal as text
}

