\docType{methods}
\name{show}
\alias{as.character}
\alias{as.character,SVGTransform-method}
\alias{print.SVGTransform}
\alias{print.SVGTransform,SVGTransform-method}
\alias{show}
\alias{show,SVGTransform-method}
\title{Show method for SVGTransform objects}
\arguments{
  \item{object}{is a \code{\link{SVGTransform}} instance.}
}
\description{
  Various methods to show or convert SVGTransform objects
}
\details{
  The \code{show()} method called with an SVGTransform
  object as an argument allows display the current
  transforms operations as a string.

  The \code{as.character()} method converts an SVGTransform
  object into a string
}

