\docType{methods}
\name{Compare}
\alias{Compare}
\alias{Compare,SVGUnit-method}
\alias{Logic}
\alias{Logic,SVGUnit-method}
\alias{Ops}
\alias{Ops,SVGUnit-method}
\title{Group Generic methods for SVGUnit objects}
\description{
  Some of the S4 group generic methods are implemented for
  SVGUnit object. For more details see
  \link{S4groupGeneric} documentation page.
}

