\docType{methods}
\name{show}
\alias{show}
\alias{show,SVGUnit-method}
\title{Show method for SVGUnit objects}
\arguments{
  \item{object}{is a \code{\link{SVGUnit}} instance.}
}
\description{
  The \code{show()} method called with an SVGUnit object as
  an argument allows display the unit value in the current
  unit as a string.
}

