\docType{methods}
\name{show}
\alias{as.character}
\alias{as.character,SVGUnit-method}
\alias{print.SVGUnit}
\alias{print.SVGUnit,SVGUnit-method}
\alias{show}
\alias{show,SVGUnit-method}
\title{Show method for SVGUnit objects}
\arguments{
  \item{object}{is a \code{\link{SVGUnit}} instance.}
}
\description{
  Various methods to show or convert SVGUnit objects
}
\details{
  The \code{show()} method called with an SVGUnit object as
  an argument allows display the unit value in the current
  unit as a string.

  The \code{as.character()} method converts an SVGUnit
  object into a string
}

